[Data] Defined core data model and implemented user storage

+++ ZGithub/Modules/Domain/Entity/TypeDefines.swift
@@ -8,4 +8,4 @@
 
 import Foundation
 
-typealias UserID = Int
+typealias UserID = Int64
